Each circle has 'includes' data bound to it
These correspond to the index/indices of clusterData which are contained
in this point. To access all the points you would use

var allPoints = [];
function(d) {
for(var i = 0; i < d.includes.length; i++){
allPoints = allPoints.concat(clusterData[d.timestep][d.includes[i]];
}
}

allPoints would contain the indices in data of all of the points in the
cluster for which "d" was bound to.